+++ Examples/java/src/test/java/software/amazon/cryptography/example/hierarchy/mutations/KmsInFlightMutationV1ToV1Test.java
@@ -11,7 +11,7 @@ import software.amazon.cryptography.example.hierarchy.AdminProvider;
 import software.amazon.cryptography.keystore.model.HierarchyVersion;
 import software.amazon.cryptography.keystoreadmin.model.Mutations;
 
-public class MutationV1ToV2Test {
+public class KmsInFlightMutationV1ToV1Test {
 
   @Test
   public void testDecryptEncryptOriginalException() {
@@ -19,7 +19,7 @@ public class MutationV1ToV2Test {
       "test-mutation-kms-access-in-flight-original-" +
       UUID.randomUUID().toString();
 
-    MutationKmsAccessInFlightTestRunner.createHappyCaseId(
+    KmsInFlightMutationAccessDeniedRunner.createHappyCaseId(
       MRK_ARN_WEST,
       branchKeyId,
       AdminProvider.admin(),
@@ -30,10 +30,9 @@ public class MutationV1ToV2Test {
       Mutations mutations = Mutations
         .builder()
         .TerminalKmsArn(POSTAL_HORN_KEY_ARN)
-        .TerminalHierarchyVersion(HierarchyVersion.v2)
         .build();
 
-      MutationKmsAccessInFlightTestRunner.runMutationTest(
+      KmsInFlightMutationAccessDeniedRunner.runMutationTest(
         branchKeyId,
         mutations,
         AdminProvider.decryptEncryptStrategy(
@@ -59,7 +58,7 @@ public class MutationV1ToV2Test {
       "test-mutation-kms-access-in-flight-terminal-" +
       UUID.randomUUID().toString();
 
-    MutationKmsAccessInFlightTestRunner.createHappyCaseId(
+    KmsInFlightMutationAccessDeniedRunner.createHappyCaseId(
       POSTAL_HORN_KEY_ARN,
       branchKeyId,
       AdminProvider.admin(),
@@ -70,10 +69,9 @@ public class MutationV1ToV2Test {
       Mutations mutations = Mutations
         .builder()
         .TerminalKmsArn(MRK_ARN_WEST)
-        .TerminalHierarchyVersion(HierarchyVersion.v2)
         .build();
 
-      MutationKmsAccessInFlightTestRunner.runMutationTest(
+      KmsInFlightMutationAccessDeniedRunner.runMutationTest(
         branchKeyId,
         mutations,
         AdminProvider.decryptEncryptStrategy(
@@ -94,12 +92,12 @@ public class MutationV1ToV2Test {
   }
 
   @Test
-  public void testKmsSimpleOriginalException() {
+  public void testReEncryptOriginalException() {
     String branchKeyId =
       "test-mutation-kms-access-in-flight-original-" +
       UUID.randomUUID().toString();
 
-    MutationKmsAccessInFlightTestRunner.createHappyCaseId(
+    KmsInFlightMutationAccessDeniedRunner.createHappyCaseId(
       MRK_ARN_WEST,
       branchKeyId,
       AdminProvider.admin(),
@@ -110,14 +108,13 @@ public class MutationV1ToV2Test {
       Mutations mutations = Mutations
         .builder()
         .TerminalKmsArn(POSTAL_HORN_KEY_ARN)
-        .TerminalHierarchyVersion(HierarchyVersion.v2)
         .build();
 
-      MutationKmsAccessInFlightTestRunner.runMutationTest(
+      KmsInFlightMutationAccessDeniedRunner.runMutationTest(
         branchKeyId,
         mutations,
-        AdminProvider.kmsSimpleStrategy(Fixtures.kmsClientWest2),
-        AdminProvider.kmsSimpleStrategy(Fixtures.denyMrkKmsClient),
+        AdminProvider.reEncryptStrategy(Fixtures.kmsClientWest2),
+        AdminProvider.reEncryptStrategy(Fixtures.denyMrkKmsClient),
         1, // expectedExceptionCount
         true, // expectFromException
         false // expectToException
@@ -128,12 +125,12 @@ public class MutationV1ToV2Test {
   }
 
   @Test
-  public void testKmsSimpleTerminalException() {
+  public void testReEncryptTerminalException() {
     String branchKeyId =
       "test-mutation-kms-access-in-flight-terminal-" +
       UUID.randomUUID().toString();
 
-    MutationKmsAccessInFlightTestRunner.createHappyCaseId(
+    KmsInFlightMutationAccessDeniedRunner.createHappyCaseId(
       POSTAL_HORN_KEY_ARN,
       branchKeyId,
       AdminProvider.admin(),
@@ -144,14 +141,13 @@ public class MutationV1ToV2Test {
       Mutations mutations = Mutations
         .builder()
         .TerminalKmsArn(MRK_ARN_WEST)
-        .TerminalHierarchyVersion(HierarchyVersion.v2)
         .build();
 
-      MutationKmsAccessInFlightTestRunner.runMutationTest(
+      KmsInFlightMutationAccessDeniedRunner.runMutationTest(
         branchKeyId,
         mutations,
-        AdminProvider.kmsSimpleStrategy(Fixtures.kmsClientWest2),
-        AdminProvider.kmsSimpleStrategy(Fixtures.denyMrkKmsClient),
+        AdminProvider.reEncryptStrategy(Fixtures.kmsClientWest2),
+        AdminProvider.reEncryptStrategy(Fixtures.denyMrkKmsClient),
         2, // expectedExceptionCount
         false, // expectFromException
         true // expectToException
