+++ Examples/java/src/test/java/software/amazon/cryptography/example/hierarchy/mutations/KmsInFlightMutationV1ToV2Test.java
@@ -11,7 +11,7 @@ import software.amazon.cryptography.example.hierarchy.AdminProvider;
 import software.amazon.cryptography.keystore.model.HierarchyVersion;
 import software.amazon.cryptography.keystoreadmin.model.Mutations;
 
-public class MutationV1ToV1Test {
+public class KmsInFlightMutationV1ToV2Test {
 
   @Test
   public void testDecryptEncryptOriginalException() {
@@ -19,7 +19,7 @@ public class MutationV1ToV1Test {
       "test-mutation-kms-access-in-flight-original-" +
       UUID.randomUUID().toString();
 
-    MutationKmsAccessInFlightTestRunner.createHappyCaseId(
+    KmsInFlightMutationAccessDeniedRunner.createHappyCaseId(
       MRK_ARN_WEST,
       branchKeyId,
       AdminProvider.admin(),
@@ -30,9 +30,10 @@ public class MutationV1ToV1Test {
       Mutations mutations = Mutations
         .builder()
         .TerminalKmsArn(POSTAL_HORN_KEY_ARN)
+        .TerminalHierarchyVersion(HierarchyVersion.v2)
         .build();
 
-      MutationKmsAccessInFlightTestRunner.runMutationTest(
+      KmsInFlightMutationAccessDeniedRunner.runMutationTest(
         branchKeyId,
         mutations,
         AdminProvider.decryptEncryptStrategy(
@@ -58,7 +59,7 @@ public class MutationV1ToV1Test {
       "test-mutation-kms-access-in-flight-terminal-" +
       UUID.randomUUID().toString();
 
-    MutationKmsAccessInFlightTestRunner.createHappyCaseId(
+    KmsInFlightMutationAccessDeniedRunner.createHappyCaseId(
       POSTAL_HORN_KEY_ARN,
       branchKeyId,
       AdminProvider.admin(),
@@ -69,9 +70,10 @@ public class MutationV1ToV1Test {
       Mutations mutations = Mutations
         .builder()
         .TerminalKmsArn(MRK_ARN_WEST)
+        .TerminalHierarchyVersion(HierarchyVersion.v2)
         .build();
 
-      MutationKmsAccessInFlightTestRunner.runMutationTest(
+      KmsInFlightMutationAccessDeniedRunner.runMutationTest(
         branchKeyId,
         mutations,
         AdminProvider.decryptEncryptStrategy(
@@ -92,12 +94,12 @@ public class MutationV1ToV1Test {
   }
 
   @Test
-  public void testReEncryptOriginalException() {
+  public void testKmsSimpleOriginalException() {
     String branchKeyId =
       "test-mutation-kms-access-in-flight-original-" +
       UUID.randomUUID().toString();
 
-    MutationKmsAccessInFlightTestRunner.createHappyCaseId(
+    KmsInFlightMutationAccessDeniedRunner.createHappyCaseId(
       MRK_ARN_WEST,
       branchKeyId,
       AdminProvider.admin(),
@@ -108,13 +110,14 @@ public class MutationV1ToV1Test {
       Mutations mutations = Mutations
         .builder()
         .TerminalKmsArn(POSTAL_HORN_KEY_ARN)
+        .TerminalHierarchyVersion(HierarchyVersion.v2)
         .build();
 
-      MutationKmsAccessInFlightTestRunner.runMutationTest(
+      KmsInFlightMutationAccessDeniedRunner.runMutationTest(
         branchKeyId,
         mutations,
-        AdminProvider.reEncryptStrategy(Fixtures.kmsClientWest2),
-        AdminProvider.reEncryptStrategy(Fixtures.denyMrkKmsClient),
+        AdminProvider.kmsSimpleStrategy(Fixtures.kmsClientWest2),
+        AdminProvider.kmsSimpleStrategy(Fixtures.denyMrkKmsClient),
         1, // expectedExceptionCount
         true, // expectFromException
         false // expectToException
@@ -125,12 +128,12 @@ public class MutationV1ToV1Test {
   }
 
   @Test
-  public void testReEncryptTerminalException() {
+  public void testKmsSimpleTerminalException() {
     String branchKeyId =
       "test-mutation-kms-access-in-flight-terminal-" +
       UUID.randomUUID().toString();
 
-    MutationKmsAccessInFlightTestRunner.createHappyCaseId(
+    KmsInFlightMutationAccessDeniedRunner.createHappyCaseId(
       POSTAL_HORN_KEY_ARN,
       branchKeyId,
       AdminProvider.admin(),
@@ -141,13 +144,14 @@ public class MutationV1ToV1Test {
       Mutations mutations = Mutations
         .builder()
         .TerminalKmsArn(MRK_ARN_WEST)
+        .TerminalHierarchyVersion(HierarchyVersion.v2)
         .build();
 
-      MutationKmsAccessInFlightTestRunner.runMutationTest(
+      KmsInFlightMutationAccessDeniedRunner.runMutationTest(
         branchKeyId,
         mutations,
-        AdminProvider.reEncryptStrategy(Fixtures.kmsClientWest2),
-        AdminProvider.reEncryptStrategy(Fixtures.denyMrkKmsClient),
+        AdminProvider.kmsSimpleStrategy(Fixtures.kmsClientWest2),
+        AdminProvider.kmsSimpleStrategy(Fixtures.denyMrkKmsClient),
         2, // expectedExceptionCount
         false, // expectFromException
         true // expectToException
